Add files via upload
A sentiment analysis project involves leveraging natural language processing techniques to discern the emotional tone or sentiment expressed in textual data. The primary objective is to gain insights into the opinions and attitudes conveyed in a given dataset, which could range from customer reviews and social media posts to survey responses. The project begins with the collection of relevant textual data, followed by meticulous preprocessing to clean and structure the information. Depending on the chosen approach, the data may be labeled with sentiment classes such as positive, negative, or neutral.